Remove noisy lines from error call stack
The callstack for a simple AssertionFailedError included a lot of lines
detailing the plumbing of OEUnit which makes it difficult to quickly
find which line number the actual error was thrown from - this is
particularly important when a lot of assertions are grouped inside a
test method. This change removes/filters the call stack of the error
objects to show calls from within the user's code - i.e. not OEUnit.